fix(#898): apply fix pass changes
